+++ src/pages/AccountSettings/tabs/Admin/Admin.spec.jsx
@@ -2,7 +2,6 @@ import { render, screen } from '@testing-library/react'
 import { MemoryRouter, Route } from 'react-router-dom'
 
 import { useUser } from 'services/user'
-import { useFlags } from 'shared/featureFlags'
 
 import Admin from './Admin'
 
@@ -12,7 +11,6 @@ jest.mock('./StudentCard', () => () => 'StudentCard')
 jest.mock('./GithubIntegrationCard', () => () => 'GithubIntegrationCard')
 jest.mock('./ManageAdminCard', () => () => 'ManageAdminCard')
 jest.mock('./DeletionCard', () => () => 'DeletionCard')
-jest.mock('shared/featureFlags')
 
 describe('AdminTab', () => {
   const defaultProps = {
@@ -20,7 +18,7 @@ describe('AdminTab', () => {
     owner: 'codecov',
   }
 
-  function setup({ owner, showThemeToggle = false }) {
+  function setup({ owner }) {
     useUser.mockReturnValue({
       data: {
         user: {
@@ -32,8 +30,6 @@ describe('AdminTab', () => {
       ...defaultProps,
       owner,
     }
-
-    useFlags.mockReturnValue({ showThemeToggle })
 
     render(
       <MemoryRouter initialEntries={['/account/gh/codecov']}>
@@ -88,37 +84,6 @@ describe('AdminTab', () => {
     it('renders the DeletionCard', () => {
       const card = screen.getByText(/DeletionCard/)
       expect(card).toBeInTheDocument()
-    })
-  })
-
-  describe('when show theme toggle flag is set to true', () => {
-    beforeEach(() => {
-      setup({ owner: 'rula', showThemeToggle: true })
-    })
-
-    it('renders colorblind label', () => {
-      const label = screen.getByText(/Colorblind Friendly/)
-      expect(label).toBeInTheDocument()
-    })
-
-    it('renders colorblind toggle', () => {
-      const button = screen.getByTestId('switch')
-      expect(button).toBeInTheDocument()
-    })
-
-    describe('on toggle switch', () => {
-      window.localStorage.__proto__.setItem = jest.fn()
-
-      beforeEach(() => {
-        screen.getByTestId('switch').click()
-      })
-
-      it('sets color-blind theme in local storage', () => {
-        expect(localStorage.setItem).toBeCalledWith(
-          'current-theme',
-          'color-blind'
-        )
-      })
     })
   })
 })
